ENG-7171 - Update LowMem Check to happen in one spot
